Move custom Liquibase changes to versioned package
So it's easier to tell which version a change class belongs to.
