+++ src/test/resources/migration/custom/RenameNumberedIndexesChangeTest-changelog.xml
@@ -9,6 +9,6 @@
             http://www.liquibase.org/xml/ns/dbchangelog
             http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-latest.xsd">
     <changeSet id="foo" author="bar">
-        <customChange class="org.dependencytrack.persistence.migration.change.RenameNumberedIndexesChange"/>
+        <customChange class="org.dependencytrack.persistence.migration.change.v530.RenameNumberedIndexesChange"/>
     </changeSet>
 </databaseChangeLog>
